Implement a timeout mechanism
This causes the simulation to simply stop. It could be done in a nicer
way where before and/or after each FMI/solver call a check if the
process has been interrupted (or perhaps is close to its timeout) and
should exit cleanly. This is a rough first attempt to just kills the
process (although destructors, etc are executed).